+++ program/source/camera.cpp
@@ -5,12 +5,11 @@
 
 #include "SDL3/SDL_scancode.h"
 #include "cse/camera.hpp"
-#include "glm/ext/vector_float3.hpp"
+#include "glm/ext/vector_double3.hpp"
 
 namespace csg
 {
-  camera::camera(const glm::vec3 &translation_)
-    : cse::camera({translation_, {0.0f, 0.0f, -1.0f}, {0.0f, 1.0f, 0.0f}}, 45.0)
+  camera::camera(const glm::dvec3 &translation_) : cse::camera({translation_, {0.0, 0.0, -1.0}, {0.0, 1.0, 0.0}}, 45.0)
   {
     hooks.set(hook::INPUT,
               [this](const bool *keys)
@@ -29,22 +28,22 @@ namespace csg
               });
 
     hooks.set(hook::SIMULATE,
-              [this](const float poll_rate)
+              [this](const double poll_rate)
               {
                 auto &position{state.active.translation.value};
                 auto &velocity{state.active.translation.rate};
                 auto &acceleration{state.active.translation.curve};
                 velocity += acceleration * poll_rate;
-                acceleration = {0.0f, 0.0f, 0.0f};
+                acceleration = {0.0, 0.0, 0.0};
                 for (int index{}; index < 3; ++index)
                 {
-                  auto drag = std::abs(velocity[index]) * (1 - (friction / max_velocity)) + friction;
-                  if (velocity[index] > 0.0f)
-                    velocity[index] = std::max(0.0f, velocity[index] - drag * poll_rate);
-                  else if (velocity[index] < -0.0f)
-                    velocity[index] = std::min(0.0f, velocity[index] + drag * poll_rate);
+                  auto drag = std::abs(velocity[index]) * (1.0 - (friction / max_velocity)) + friction;
+                  if (velocity[index] > 0.0)
+                    velocity[index] = std::max(0.0, velocity[index] - drag * poll_rate);
+                  else if (velocity[index] < -0.0)
+                    velocity[index] = std::min(0.0, velocity[index] + drag * poll_rate);
                   else
-                    velocity[index] = 0.0f;
+                    velocity[index] = 0.0;
                 }
                 position += velocity * poll_rate;
